cli: align with sealed event runtime APIs
- replace direct event envelope construction with typed constructors

- read sealed draft and event data through accessors

- update proxy publish tests for raw signed event JSON requests

- preserve u64 event timestamps in CLI output views

+++ src/view/runtime.rs
@@ -3036,7 +3036,7 @@ pub struct ListingMutationEventView {
     pub kind: u32,
     pub author: String,
     #[serde(skip_serializing_if = "Option::is_none")]
-    pub created_at: Option<u32>,
+    pub created_at: Option<u64>,
     pub content: String,
     pub tags: Vec<Vec<String>>,
     #[serde(skip_serializing_if = "Option::is_none")]
